requirements.txt
streamlit
pdfminer.six
sentence-transformers
scikit-learn
python-dotenv
groq
reportlab
google-generativeai
python-dotenv
